requirements.txt
configargparse
jinja2
pyyaml
requests
simplejson
